Uninitialize peer when client is terminated by debugger
There is a logic in code which skips peer uninitialization if
it was triggered by different process which has added peer.
This logic was added to handle case when EVT_WDF_FILE_CLEANUP is
mysteriously called on driver reinstall, when no peer added yet.

When client is terminated by debugger, EVT_WDF_FILE_CLEANUP callback
is called from the context of debugger process, which differs from the process which
initiated OVPN_IOCTL_NEW_PEER, which triggers logic above.

Fix by removing pid comparison and skip uninitialization only if pid is 0.

Reported-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: Lev Stipakov <lev@openvpn.net>