Drop DMF and use pre-allocated buffer pools
DMF is fairly large compile-time dependency, from which
we use a very small part. Besides we have own patches on top
of it, which complicates maintenance.

This commit implements:

 - buffer pool with pre-allocated buffers of fixed size
 - queue

Both implementaions are based on doubly-linked
list and use interlocked functions for atomic modifications.

On TX datapath, we fetch buffer from TX pool and pass it to
completion routine, where we return buffer back to the TX pool.

On RX datapath, we fetch buffer from RX pool and enqueue it into
control or data queue. When we got read request from userspace,
we dequeue buffer from control queue and put it back to RX pool.
When NetAdapter calls us to indicate packets to OS, we dequeue buffer
from data queue and put it back to RX pool.

When userspace closes the file handle (also happens when app exits),
we flush control queue so that new client won't receive control channel
messages from the previous session.

Signed-off-by: Lev Stipakov <lev@openvpn.net>