GitHub Actions: add Debug build configuration
In order to install driver on the machine with test
signing enabled, the driver itself needs to be test signed.

For release builds we do signing manually, so we disabled it
in build. To allow people to just grab driver from GitHub Actions
artifacts, add debug configuration, which signs driver with a
test certficatie.

While on it, put Win32 artifacts under its own arch directory,
same as we do for arm64 and x64.

Signed-off-by: Lev Stipakov <lev@openvpn.net>