Change adapter creation lifecycle
Create/start adapter on driver initialization
and stop/destroy on cleanup callback. On creation,
set connection status to "disconnected" and
on "start_vpn" ioctl set to "connected".

Besides that this lifecycle is recommended by MSFT

https://docs.microsoft.com/en-us/windows-hardware/drivers/netcx/device-and-adapter-initialization

this also make dco-win adapter visible in Wireshark.
Without this change, it is visible only when connected,
which could be confusing, since other adapters
are always visible.

Signed-off-by: Lev Stipakov <lev@openvpn.net>