Add support for persistent tunnels
In some cases we want to prevent traffic leakage
on reconnect. In this case we need to teardown
VPN connection without stopping tun adapter.
This is achieved in openvpn with --persist-tun option.

Add DEL_PEER command which clears
peer state - destroys timers, uninits crypto,
closes socket, flushes control packet buffers
and finishes pending read requests. It doesn't change
adapter state - it stays connected so that routes
are kept intact.

Signed-off-by: Lev Stipakov <lev@openvpn.net>