Set adapter state to "disconnected" on file close even if peer is removed
After peer is removed we don't immediately set adapter state
to "disconnected" in order to support persist tun. We are supposed to
do it after the last handle to the device is closed. Due to bug in code,
in file close callback we just exit if peer deletion failed
(even though it is normal il peer was explicitly deleted by client)
and skipped changing adapter state.

Instead of exiting, ignore peer removal result (it is logged already)
and proceed to changing adapter state.

Signed-off-by: Lev Stipakov <lev@openvpn.net>