Improve pool memory management
 - Handle case when IoAllocateMdl() and ExAllocatePool2()
 return NULL

 - Set max pool size (in-flight packets) to 100'000.

This might help with https://github.com/OpenVPN/ovpn-dco-win/issues/47

By playing with Quick Edit in console window (which makes execution
stop) where iPerf3 was running I was able to get pool size growing
to unreasonable values (over 50k). My understanding is that normally it
should be <10k, but let's be on a safe side and cap it to 100k - in
kernel dump it was over 1mil and system ran out of memory.

Bump to 0.9.4.

Signed-off-by: Lev Stipakov <lev@openvpn.net>