Fix bug check in timer management routines
Commit b750b2c ("timer: refactor timers implementation")
has changed timer initialization logic and introduced a bug,
where calling SET_PEER ioctl after failed NEW_PEER ioctl
causes WDF_VIOLATION bug check, because we try to get a context
of a WDF object which is NULL.

Fix by adding NULL checks.

Note that this is not expected to happen with openvpn as a
driver client.

CVE: 2024-5198

Reported-By: Lukas Jokubauskas <lukas.jokubauskas@nordsec.com>
Signed-off-by: Lev Stipakov <lev@openvpn.net>