Ensure that TCP flag is properly reset
When stopping VPN, we must clear TCP
flag of socket state. When starting VPN
in MP mode, we set TCP flag to FALSE, since
we don't support TCP is MP mode.

Without that, if we first start VPN as a client
in TCP mode and then disconnect and start
in MP mode, we got bug check inside WSK
since we attempt to use TCP API on UDP socket.

Bump to 2.5.3.

Signed-off-by: Lev Stipakov <lev@openvpn.net>