P2P: Restore replacement behavior for existing peers
Commit

  06099dc ("Move peer-specific data to separate struct")

introduced a behavior change in the NEW_PEER ioctl,
causing it to fail if a peer already exists. Previously,
the behavior was to delete the old peer before adding the new one.

This issue can occur when a client attempts to connect to
an unreachable TCP server. In such cases, the peer is added,
but the subsequent connection attempt (still part of the ioctl) fails.
The client then retries by sending the NEW_PEER ioctl again,
but this fails because the peer was already added.

Fix this by restoring the previous behavior,
removing the existing peer before adding a new one.

Bump version to 2.5.4.

Signed-off-by: Lev Stipakov <lev@openvpn.net>